+++ scripts/polyfills.js
@@ -3,7 +3,6 @@
  * @fileoverview Polyfills for old browsers.
  * @see https://github.com/inexorabletash/polyfill/blob/master/es5.js
  * @see https://gist.github.com/jhermsmeier/9a34b06a107bbf5d2c91
- * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/codePointAt
  */
 
 // ES 15.2.3.6 Object.defineProperty ( O, P, Attributes )
@@ -35,14 +34,9 @@
 // called with DOM elements; Here it is tested against a non-DOM object.
 // If an error is raised, the method is replaced.
 // https://gist.github.com/jhermsmeier/9a34b06a107bbf5d2c91
-var testObject = {"t":"o"};
-var replaceGetOwnPropertyDescriptor = false;
 try {
-  Object.getOwnPropertyDescriptor(testObject, "t");
+  Object.getOwnPropertyDescriptor({"t":"o"}, "t");
 } catch(err) {
-  replaceGetOwnPropertyDescriptor = true;
-}
-if (replaceGetOwnPropertyDescriptor) {
   Object.getOwnPropertyDescriptor = function( object, key ) {
     
     var hasSupport =
@@ -66,60 +60,4 @@ if (replaceGetOwnPropertyDescriptor) {
       value: object[ key ]
     }
   }
-}
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/codePointAt
-/*! https://mths.be/codepointat v0.2.0 by @mathias */
-if (!String.prototype.codePointAt) {
-  (function() {
-    'use strict'; // needed to support `apply`/`call` with `undefined`/`null`
-    var defineProperty = (function() {
-      // IE 8 only supports `Object.defineProperty` on DOM elements
-      try {
-        var object = {};
-        var $defineProperty = Object.defineProperty;
-        var result = $defineProperty(object, object, object) && $defineProperty;
-      } catch(error) {}
-      return result;
-    }());
-    var codePointAt = function(position) {
-      if (this == null) {
-        throw TypeError();
-      }
-      var string = String(this);
-      var size = string.length;
-      // `ToInteger`
-      var index = position ? Number(position) : 0;
-      if (index != index) { // better `isNaN`
-        index = 0;
-      }
-      // Account for out-of-bounds indices:
-      if (index < 0 || index >= size) {
-        return undefined;
-      }
-      // Get the first code unit
-      var first = string.charCodeAt(index);
-      var second;
-      if ( // check if it’s the start of a surrogate pair
-        first >= 0xD800 && first <= 0xDBFF && // high surrogate
-        size > index + 1 // there is a next code unit
-      ) {
-        second = string.charCodeAt(index + 1);
-        if (second >= 0xDC00 && second <= 0xDFFF) { // low surrogate
-          // https://mathiasbynens.be/notes/javascript-encoding#surrogate-formulae
-          return (first - 0xD800) * 0x400 + second - 0xDC00 + 0x10000;
-        }
-      }
-      return first;
-    };
-    if (defineProperty) {
-      defineProperty(String.prototype, 'codePointAt', {
-        'value': codePointAt,
-        'configurable': true,
-        'writable': true
-      });
-    } else {
-      String.prototype.codePointAt = codePointAt;
-    }
-  }());
 }